Add reference to original network monitor script
Just pointing to the gist that holds the shell version of the monitor. This should help any one that can use a terminal not have to deal with the objective c + apple side of things.
Also, for unix users that link 'say' and 'terminal-notifier' (or change the commands in the script).